Fixes to vector type for nvcc
