+++ opm/material/fluidsystems/BlackOilFluidSystem.cpp
@@ -39,8 +39,8 @@
 namespace Opm {
 
 #if HAVE_ECL_INPUT
-template <class Scalar, class IndexTraits>
-void BlackOilFluidSystem<Scalar,IndexTraits>::
+template <class Scalar, class IndexTraits, template<typename> typename Storage>
+void BlackOilFluidSystem<Scalar,IndexTraits, Storage>::
 initFromState(const EclipseState& eclState, const Schedule& schedule)
 {
     if (eclState.getSimulationConfig().useEnthalpy()) {
@@ -224,213 +224,38 @@ initFromState(const EclipseState& eclState, const Schedule& schedule)
 }
 #endif
 
-template <class Scalar, class IndexTraits>
-void BlackOilFluidSystem<Scalar,IndexTraits>::
-initBegin(std::size_t numPvtRegions)
-{
-    isInitialized_ = false;
-    useSaturatedTables_ = true;
-
-    enableDissolvedGas_ = true;
-    enableDissolvedGasInWater_ = false;
-    enableVaporizedOil_ = false;
-    enableVaporizedWater_ = false;
-    enableDiffusion_ = false;
-
-    oilPvt_ = nullptr;
-    gasPvt_ = nullptr;
-    waterPvt_ = nullptr;
-
-    surfaceTemperature = 273.15 + 15.56; // [K]
-    surfacePressure = 1.01325e5; // [Pa]
-    setReservoirTemperature(surfaceTemperature);
-
-    numActivePhases_ = numPhases;
-    std::fill_n(&phaseIsActive_[0], numPhases, true);
-
-    resizeArrays_(numPvtRegions);
-}
-
-template <class Scalar, class IndexTraits>
-void BlackOilFluidSystem<Scalar,IndexTraits>::
-setReferenceDensities(Scalar rhoOil,
-                      Scalar rhoWater,
-                      Scalar rhoGas,
-                      unsigned regionIdx)
-{
-    referenceDensity_[regionIdx][oilPhaseIdx] = rhoOil;
-    referenceDensity_[regionIdx][waterPhaseIdx] = rhoWater;
-    referenceDensity_[regionIdx][gasPhaseIdx] = rhoGas;
-}
-
-template <class Scalar, class IndexTraits>
-void BlackOilFluidSystem<Scalar,IndexTraits>::initEnd()
-{
-    // calculate the final 2D functions which are used for interpolation.
-    const std::size_t num_regions = molarMass_.size();
-    for (unsigned regionIdx = 0; regionIdx < num_regions; ++regionIdx) {
-        // calculate molar masses
-
-        // water is simple: 18 g/mol
-        molarMass_[regionIdx][waterCompIdx] = 18e-3;
-
-        if (phaseIsActive(gasPhaseIdx)) {
-            // for gas, we take the density at standard conditions and assume it to be ideal
-            Scalar p = surfacePressure;
-            Scalar T = surfaceTemperature;
-            Scalar rho_g = referenceDensity_[/*regionIdx=*/0][gasPhaseIdx];
-            molarMass_[regionIdx][gasCompIdx] = Constants<Scalar>::R*T*rho_g / p;
-        }
-        else
-            // hydrogen gas. we just set this do avoid NaNs later
-            molarMass_[regionIdx][gasCompIdx] = 2e-3;
-
-        // finally, for oil phase, we take the molar mass from the spe9 paper
-        molarMass_[regionIdx][oilCompIdx] = 175e-3; // kg/mol
-    }
-
-
-    int activePhaseIdx = 0;
-    for (unsigned phaseIdx = 0; phaseIdx < numPhases; ++phaseIdx) {
-        if(phaseIsActive(phaseIdx)){
-            canonicalToActivePhaseIdx_[phaseIdx] = activePhaseIdx;
-            activeToCanonicalPhaseIdx_[activePhaseIdx] = phaseIdx;
-            activePhaseIdx++;
-        }
-    }
-    isInitialized_ = true;
-}
-
-template <class Scalar, class IndexTraits>
-std::string_view BlackOilFluidSystem<Scalar,IndexTraits>::
-phaseName(unsigned phaseIdx)
-{
-    switch (phaseIdx) {
-    case waterPhaseIdx:
-        return "water";
-    case oilPhaseIdx:
-        return "oil";
-    case gasPhaseIdx:
-        return "gas";
-
-    default:
-        throw std::logic_error(fmt::format("Phase index {} is unknown", phaseIdx));
-    }
-}
-
-template <class Scalar, class IndexTraits>
-unsigned BlackOilFluidSystem<Scalar,IndexTraits>::
-solventComponentIndex(unsigned phaseIdx)
-{
-    switch (phaseIdx) {
-    case waterPhaseIdx:
-        return waterCompIdx;
-    case oilPhaseIdx:
-        return oilCompIdx;
-    case gasPhaseIdx:
-        return gasCompIdx;
-
-    default:
-        throw std::logic_error(fmt::format("Phase index {} is unknown", phaseIdx));
-    }
-}
-
-template <class Scalar, class IndexTraits>
-unsigned BlackOilFluidSystem<Scalar,IndexTraits>::
-soluteComponentIndex(unsigned phaseIdx)
-{
-    switch (phaseIdx) {
-    case waterPhaseIdx:
-        if (enableDissolvedGasInWater())
-            return gasCompIdx;
-        throw std::logic_error("The water phase does not have any solutes in the black oil model!");
-    case oilPhaseIdx:
-        return gasCompIdx;
-    case gasPhaseIdx:
-        if (enableVaporizedWater()) {
-            return waterCompIdx;
-        }
-        return oilCompIdx;
-
-    default:
-        throw std::logic_error(fmt::format("Phase index {} is unknown", phaseIdx));
-    }
-}
-
-template <class Scalar, class IndexTraits>
-std::string_view BlackOilFluidSystem<Scalar,IndexTraits>::
-componentName(unsigned compIdx)
-{
-    switch (compIdx) {
-    case waterCompIdx:
-        return "Water";
-    case oilCompIdx:
-        return "Oil";
-    case gasCompIdx:
-        return "Gas";
-
-    default:
-        throw std::logic_error(fmt::format("Component index {} is unknown", compIdx));
-    }
-}
-
-template <class Scalar, class IndexTraits>
-short BlackOilFluidSystem<Scalar,IndexTraits>::
-activeToCanonicalPhaseIdx(unsigned activePhaseIdx)
-{
-    assert(activePhaseIdx<numActivePhases());
-    return activeToCanonicalPhaseIdx_[activePhaseIdx];
-}
-
-template <class Scalar, class IndexTraits>
-short BlackOilFluidSystem<Scalar,IndexTraits>::
-canonicalToActivePhaseIdx(unsigned phaseIdx)
-{
-    assert(phaseIdx<numPhases);
-    assert(phaseIsActive(phaseIdx));
-    return canonicalToActivePhaseIdx_[phaseIdx];
-}
-
-template <class Scalar, class IndexTraits>
-void BlackOilFluidSystem<Scalar,IndexTraits>::
-resizeArrays_(std::size_t numRegions)
-{
-    molarMass_.resize(numRegions);
-    referenceDensity_.resize(numRegions);
-}
-
 #define INSTANTIATE_TYPE(T) \
-    template<> unsigned char BlackOilFluidSystem<T, BlackOilDefaultIndexTraits>::numActivePhases_ = 0; \
-    template<> std::array<bool, BlackOilFluidSystem<T, BlackOilDefaultIndexTraits>::numPhases> \
-        BlackOilFluidSystem<T, BlackOilDefaultIndexTraits>::phaseIsActive_ = {false, false, false}; \
-    template<> std::array<short, BlackOilFluidSystem<T, BlackOilDefaultIndexTraits>::numPhases> \
-        BlackOilFluidSystem<T, BlackOilDefaultIndexTraits>::activeToCanonicalPhaseIdx_ = {0, 1, 2}; \
-    template<> std::array<short, BlackOilFluidSystem<T, BlackOilDefaultIndexTraits>::numPhases> \
-        BlackOilFluidSystem<T , BlackOilDefaultIndexTraits>::canonicalToActivePhaseIdx_ = {0, 1, 2}; \
-    template<> T BlackOilFluidSystem<T, BlackOilDefaultIndexTraits>::surfaceTemperature = 0.0; \
-    template<> T BlackOilFluidSystem<T, BlackOilDefaultIndexTraits>::surfacePressure = 0.0; \
-    template<> T BlackOilFluidSystem<T, BlackOilDefaultIndexTraits>::reservoirTemperature_ = 0.0; \
-    template<> bool BlackOilFluidSystem<T, BlackOilDefaultIndexTraits>::enableDissolvedGas_ = true; \
-    template<> bool BlackOilFluidSystem<T, BlackOilDefaultIndexTraits>::enableDissolvedGasInWater_ = false; \
-    template<> bool BlackOilFluidSystem<T, BlackOilDefaultIndexTraits>::enableVaporizedOil_ = false; \
-    template<> bool BlackOilFluidSystem<T, BlackOilDefaultIndexTraits>::enableVaporizedWater_ = false; \
-    template<> bool BlackOilFluidSystem<T, BlackOilDefaultIndexTraits>::enableDiffusion_ = false; \
+    template<> unsigned char BlackOilFluidSystem<T, BlackOilDefaultIndexTraits, VectorWithDefaultAllocator>::numActivePhases_ = 0; \
+    template<> std::array<bool, BlackOilFluidSystem<T, BlackOilDefaultIndexTraits, VectorWithDefaultAllocator>::numPhases> \
+        BlackOilFluidSystem<T, BlackOilDefaultIndexTraits, VectorWithDefaultAllocator>::phaseIsActive_ = {false, false, false}; \
+    template<> std::array<short, BlackOilFluidSystem<T, BlackOilDefaultIndexTraits, VectorWithDefaultAllocator>::numPhases> \
+        BlackOilFluidSystem<T, BlackOilDefaultIndexTraits, VectorWithDefaultAllocator>::activeToCanonicalPhaseIdx_ = {0, 1, 2}; \
+    template<> std::array<short, BlackOilFluidSystem<T, BlackOilDefaultIndexTraits, VectorWithDefaultAllocator>::numPhases> \
+        BlackOilFluidSystem<T , BlackOilDefaultIndexTraits, VectorWithDefaultAllocator>::canonicalToActivePhaseIdx_ = {0, 1, 2}; \
+    template<> T BlackOilFluidSystem<T, BlackOilDefaultIndexTraits, VectorWithDefaultAllocator>::surfaceTemperature = 0.0; \
+    template<> T BlackOilFluidSystem<T, BlackOilDefaultIndexTraits, VectorWithDefaultAllocator>::surfacePressure = 0.0; \
+    template<> T BlackOilFluidSystem<T, BlackOilDefaultIndexTraits, VectorWithDefaultAllocator>::reservoirTemperature_ = 0.0; \
+    template<> bool BlackOilFluidSystem<T, BlackOilDefaultIndexTraits, VectorWithDefaultAllocator>::enableDissolvedGas_ = true; \
+    template<> bool BlackOilFluidSystem<T, BlackOilDefaultIndexTraits, VectorWithDefaultAllocator>::enableDissolvedGasInWater_ = false; \
+    template<> bool BlackOilFluidSystem<T, BlackOilDefaultIndexTraits, VectorWithDefaultAllocator>::enableVaporizedOil_ = false; \
+    template<> bool BlackOilFluidSystem<T, BlackOilDefaultIndexTraits, VectorWithDefaultAllocator>::enableVaporizedWater_ = false; \
+    template<> bool BlackOilFluidSystem<T, BlackOilDefaultIndexTraits, VectorWithDefaultAllocator>::enableDiffusion_ = false; \
     template<> std::shared_ptr<OilPvtMultiplexer<T>> \
-        BlackOilFluidSystem<T, BlackOilDefaultIndexTraits>::oilPvt_ = {}; \
+        BlackOilFluidSystem<T, BlackOilDefaultIndexTraits, VectorWithDefaultAllocator>::oilPvt_ = {}; \
     template<> std::shared_ptr<GasPvtMultiplexer<T>> \
-        BlackOilFluidSystem<T, BlackOilDefaultIndexTraits>::gasPvt_ = {}; \
+        BlackOilFluidSystem<T, BlackOilDefaultIndexTraits, VectorWithDefaultAllocator>::gasPvt_ = {}; \
     template<> std::shared_ptr<WaterPvtMultiplexer<T>> \
-        BlackOilFluidSystem<T, BlackOilDefaultIndexTraits>::waterPvt_ = {}; \
+        BlackOilFluidSystem<T, BlackOilDefaultIndexTraits, VectorWithDefaultAllocator>::waterPvt_ = {}; \
     template<> std::vector<std::array<T, 3>> \
-        BlackOilFluidSystem<T, BlackOilDefaultIndexTraits>::referenceDensity_ = {}; \
+        BlackOilFluidSystem<T, BlackOilDefaultIndexTraits, VectorWithDefaultAllocator>::referenceDensity_ = {}; \
     template<> std::vector<std::array<T, 3>> \
-        BlackOilFluidSystem<T, BlackOilDefaultIndexTraits>::molarMass_ = {}; \
+        BlackOilFluidSystem<T, BlackOilDefaultIndexTraits, VectorWithDefaultAllocator>::molarMass_ = {}; \
     template<> std::vector<std::array<T, 9>> \
-        BlackOilFluidSystem<T, BlackOilDefaultIndexTraits>::diffusionCoefficients_ = {}; \
-    template<> bool BlackOilFluidSystem<T, BlackOilDefaultIndexTraits>::isInitialized_ = false; \
-    template<> bool BlackOilFluidSystem<T, BlackOilDefaultIndexTraits>::useSaturatedTables_ = false; \
-    template<> bool BlackOilFluidSystem<T, BlackOilDefaultIndexTraits>::enthalpy_eq_energy_ = false; \
-    template class BlackOilFluidSystem<T,BlackOilDefaultIndexTraits>;
+        BlackOilFluidSystem<T, BlackOilDefaultIndexTraits, VectorWithDefaultAllocator>::diffusionCoefficients_ = {}; \
+    template<> bool BlackOilFluidSystem<T, BlackOilDefaultIndexTraits, VectorWithDefaultAllocator>::isInitialized_ = false; \
+    template<> bool BlackOilFluidSystem<T, BlackOilDefaultIndexTraits, VectorWithDefaultAllocator>::useSaturatedTables_ = false; \
+    template<> bool BlackOilFluidSystem<T, BlackOilDefaultIndexTraits, VectorWithDefaultAllocator>::enthalpy_eq_energy_ = false; \
+    template class BlackOilFluidSystem<T,BlackOilDefaultIndexTraits, VectorWithDefaultAllocator>;
     // IMPORTANT: The class must be instantiated after the template template specializations
     //    or else the static variable above will appear as undefined in the generated object file.
